pyproject.toml
[project]
name = "echosafe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colored>=2.3.1",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pyserial>=3.5",
    "scikit-learn>=1.8.0",
    "scipy>=1.16.3",
    "tensorflow>=2.20.0",
]
